requirements.txt
requests==2.28.1
pandas==1.5.3
psycopg2==2.9.3
geopy==2.2.0
python-dotenv==0.21.1
SQLAlchemy==2.0.12
